WIP!
Combined z, trained a Gfil to imitate Gfil*z(t) = G*z(t-1).
However the thing works fine if you iterate only once, however if you
iterate more than that, it becomes very bad.

Also, to test / see, replace testPatt with trainPatt, it should be stable
(as it is with cascaded setup), and not go haywire!