fail: zold update using GinvWin
I thought it might be useful to have a cleaner state z uses successively
as the zold. However, we can not use zs{1} as this state (although it
seems to give better results), because it will be one time step ahead.

So, I used tried to find a state z' such that:
G*z'(t-1) + Win*p(t) = G*z(t-1) + Win*y(t)

ie, the net combination obtained by a cleaner signal y(t) and a worse
state z(t-1) (same as zold), to be imitated by a better state z'(t-1)
and a dirtier signal p(t). This hopes that Gz' would be able to cancel
some of the noise in p itself.