Working: adjusted lrr and mismatchexp
Lrr = 0.1 and NNets = 13 give a good result with this code.
I reduced the mismatchexp to 0.5 (which makes sense) since the higher
valuse were making the thing explode.

The autocorr errors are still 10 times worse than in cascade, and enrgy
errors are also 3-4 times worse.
Also it takes 13 iterations as opposed to only 5 cascades.

Next thing to try would be to optimise lrr and mismexp further to see how far
this can go. Then look for some way to reduce iterations.