Found out that conceptor will "force" the patt
So irrespective of what input pattern you give, as long as there is a
suff. amount of activations in the network, the conceptor (and scaling)
will force it to become the required pattern.

So, starting from Different sines, or even rnadom noise, we are able to
"filter" the signal to become closer to the original one.

Really troubling.