Final commit
The de-slanting that I'd hoped would happen does not happen at all. In
fact it is a very very difficult problem I realize, because of the
relative time shift with the inputs that arrises because of slanting,
which lead to a completely different state dynamics as opposed to the
straight one.

Also, there is no (easy, that I can think of) way to "combine" slant nd
normal pattern conceptors to negate the rel. time shift.. :/